DB_Tool_OpenClaw/postgres_mcp_server/requirements.txt
mcp
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic
pandas
